+++ lending/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.7.71",
-  genVersion: "2.502.0",
-  userAgent: "speakeasy-sdk/typescript 0.7.71 2.502.0 1.0.0 ryan-lending",
+  sdkVersion: "0.7.72",
+  genVersion: "2.503.2",
+  userAgent: "speakeasy-sdk/typescript 0.7.72 2.503.2 1.0.0 ryan-lending",
 } as const;